Use quotes
Originally I had omitted the quotes here intentionally for the
purpose of increasing the contrast between this code block and the
next.  However, it strikes me that this mis-quoting is an extremely
common unintentional source of problems and if I leave it here I
am contributing to the proliferation of bad code in the world.